Doc: add doc to state that the --log option is deprecated
Since kernel(>=v5.10), a new lockless ringbuffer is introduced. And
crash commit a5531b24 ("printk: add support for lockless ringbuffer")
implemented the lockless ringbuffer dumping for cmd_log, this however,
relies on the existence of kernel debuginfo.

Since there is already a similar function been implemented for
makedumpfile, aka "makedumpfile --dump-dmesg" to dump dmesg logs with
only vmcore, there is no need to maintain a similar code in crash
as well. In addition, this option is not widely used, so just state the
"--log" option is deprecated.

Signed-off-by: Tao Liu <ltao@redhat.com>