CI: enable ci-test with Testing Farm as github action
Currently, only enable the test on Fedora Rawhide x86 64, for other
architectures such as aarch64, ppc64le and s390x, they will fail because
of Beaker issues, not involving in Testing Farm.

Once known issues are resolved, it's easy to enable on other arches as
below:

  arch: [x86_64, aarch64, ppc64le, s390x]

Similarly, it's also easy to enable on other OS such as CentOS:

  compose: [Fedora-Rawhide, CentOS-Stream-10]

Note: if the current testcases need to be added or modified in the
future, only need to do that in the github Repo(see the Link).

Link: https://github.com/crash-utility/crash-test
Signed-off-by: Lianbo Jiang <lijiang@redhat.com>