gdb: Fix a regression for eppic extension on gdb-16.2
There is a regression found when testing eppic extension on gdb-16.2
crash:

   crash> cgroup
   /root/.eppic/cgroup.c : line 99 : Error: undefined variable 'cgroup_roots'

The root cause is when doing gdb upgrading, the replacement of
gdb_get_datatype() is incorrect:

The original gdb-10.2 version:

   long value = SYMBOL_VALUE(expr->elts[2].symbol);

The incorrect gdb-16.2 replacement:

   long value = value_as_long(expr->evaluate());

According to gdb/tracepoint.c, the correct gdb-16.2 replacement should be:

   symbol *sym;
   expr::var_value_operation *vvop
     = (gdb::checked_static_cast<expr::var_value_operation *>
        (exp->op.get ()));
   sym = vvop->get_symbol ();
   long value = sym->value_longest ();

Otherwise, the value_as_long() will throw an exception when trying to
convert a struct into long, such as "cgroup_roots". The reason why this
issue only observed on crash extensions, is the faulty code block
triggered with "req->tcb", which is a callback for gdb_interface(), and
the callback is used by eppic extension, but the normal crash internal calls
hardly use it.

After:
   crash> cgroup
   0:/user.slice/user-1000.slice/session-2.scope

Signed-off-by: Tao Liu <ltao@redhat.com>