gdb: set req->typecode when type resolvation success
In gdb_get_datatype(), the req->typecode is set as TYPE_CODE_UNDEF, and
it won't be set to other value in OP_TYPE case. If the type doesn't exist,
it will bail out in parse_expression() by exception catch of gdb_command_funnel(),
and req->typecode stays TYPE_CODE_UNDEF, which is expected. But if the type is
exist, it will be resolved by drillDownType(), so req->typecode shouldn't be
TYPE_CODE_UNDEF anymore, otherwise upper functions will regard this as a type
resolvation failure.

This issue only affects eppic extensions because currently the only user
of req->tcb is eppic.

Signed-off-by: Tao Liu <ltao@redhat.com>